Former-commit-id: 12c32e56288645c68e3312848970f277271e49e6

+++ tests/SpiderMonkeyTest/js_manual_conversions.mm
@@ -187,7 +187,7 @@ void * jsval_to_opaque( JSContext *cx, jsval vp )
 #ifdef __LP64__
 	JSObject *tmp_arg;
 	JS_ValueToObject( cx, vp, &tmp_arg );
-	NSCAssert( JS_GetTypedArrayByteLength( tmp_arg ) == sizeof(int), @"Invalid Typed Array lenght");
+	NSCAssert( JS_GetTypedArrayByteLength( tmp_arg ) == sizeof(long), @"Invalid Typed Array lenght");
 	
 	int32_t* arg_array = (int32_t*)JS_GetTypedArrayData( tmp_arg );
 	uint64 ret =  arg_array[0];
